Check if release notes submitted by Create view, and assign None if not
If Release Notes feature is disabled, the release_notes value will be
None. This way, the release_notes variable can be passed in as a parameter
regardless of whether the feature is enabled or not.

The post_release_notes function will then check for a release_notes value
and use an empty string if it is None.

This saves passing the state values through to the post_release_notes function.